notrickszone_scraper/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
python-dateutil>=2.8.0
pyyaml>=6.0
tqdm>=4.65.0
urllib3>=1.26.0

# Optional for better tokenization:
# spacy>=3.5.0
# nltk>=3.8.0
